AddEntry files Organised into Folders
