+++ client/src/App.js
@@ -1,9 +1,9 @@
 import { BrowserRouter as Router, Routes, Route } from "react-router-dom";
 import Box from "@mui/material/Box";
-import Home from "./pages/Home";
+import Home from "./pages/Home/Home";
 import Navbar from "./Component/Navbar/Navbar";
 import Footer from "./Component/Footer/Footer";
-import AddEntry from "./pages/AddEntry";
+import AddEntry from "./pages/AddEntry/AddEntry";
 import SignIn from "./pages/SignIn";
 import SignUp from "./pages/SignUp";
 import ViewEntries from "./pages/ViewEntries";
