+++ client/src/pages/AddEntry/addEntry.css
@@ -5,7 +5,7 @@
 }
 
 /* Entry Form */
-.main {
+.add-star-main {
 	width: 65%;
 	margin: 0 auto;
 	margin-top: 2rem;
